Split streaming sender out into separate build unit
